Update EstatisticaResponse.java to use BigDecimal for the sum, avg, min, and max fields

+++ src/main/java/com/leonardof108/TransactionStatsAPI/api/dto/EstatisticaResponse.java
@@ -1,9 +1,11 @@
 package com.leonardof108.TransactionStatsAPI.api.dto;
 
+import java.math.BigDecimal;
+
 public record EstatisticaResponse(
         long count,
-        double sum,
-        double avg,
-        double min,
-        double max
+        BigDecimal sum,
+        BigDecimal avg,
+        BigDecimal min,
+        BigDecimal max
 ) {}
